gpu: factor shared helpers into mintpy.gpu._common
Both mintpy.gpu.ifgram_inversion and mintpy.gpu.dem_error reduce a
per-pixel weighted least-squares problem to batched normal equations
plus cuSolver-batched Cholesky. They duplicated the torch/CUDA probe,
SUPPORTED_SOLVERS, the VRAM-aware chunk sizer, and the Cholesky core
with rank-deficient fallback. Move those into _common.py and import
them from each solver. ifgram_inversion._solve_cholesky is reduced to
a thin wrapper that assembles the weighted system; dem_error's
_solve_cholesky_batched is replaced with a direct call to the shared
core. Public APIs (estimate_timeseries_batch, estimate_dem_error_
pixelwise_batch, is_solver_available) are unchanged.

Verified by running tests/test_gpu_ifgram_inversion.py (6) and
tests/test_gpu_dem_error.py (4): 10/10 pass; the synthetic CPU vs GPU
diffs for dem_error are unchanged (delta_z rms/|cpu|max ~ 3e-6,
ts_cor ~ 6e-9, ts_res ~ 7e-5).
