+++ src/mintpy/gpu/ifgram_inversion.py
@@ -26,79 +26,26 @@ the numerical reference.
 
 import numpy as np
 
-try:
+from mintpy.gpu._common import (
+    HAS_TORCH,
+    SUPPORTED_SOLVERS,
+    auto_chunk_size,
+    get_torch_device,
+    is_solver_available,
+    solve_normal_equations_batched,
+)
+
+if HAS_TORCH:
     import torch
-    HAS_TORCH = True
-except ImportError:
-    HAS_TORCH = False
-
-
-SUPPORTED_SOLVERS = ('cpu', 'torch')
-
-# default chunk size when caller does not provide one and VRAM probing fails
-DEFAULT_CHUNK_SIZE = 20000
-
-# safety factor applied to free VRAM when auto-sizing the chunk
-VRAM_SAFETY = 0.4
-
-
-def is_solver_available(solver):
-    """Return True if the named WLS solver is importable and usable."""
-    if solver == 'cpu':
-        return True
-    if solver == 'torch':
-        return HAS_TORCH and torch.cuda.is_available()
-    return False
-
-
-def _get_torch_device(solver):
-    if not HAS_TORCH:
-        raise ImportError(
-            f"solver='{solver}' requires PyTorch. "
-            "Install with `pip install -e \".[gpu]\" "
-            "--extra-index-url https://download.pytorch.org/whl/cu128 "
-            "--index-strategy unsafe-best-match`."
-        )
-    if not torch.cuda.is_available():
-        raise RuntimeError(
-            f"solver='{solver}' requires CUDA, but torch.cuda.is_available() is False."
-        )
-    return torch.device('cuda')
-
-
-def _auto_chunk_size(num_pair, num_unknown, dtype_bytes=4):
-    """Pick chunk size from free VRAM.
-
-    The dominant per-pixel allocations during one chunk are:
-        Gw (n, num_pair, num_unknown)         ~ num_pair * num_unknown * dtype_bytes
-        N, L (n, num_unknown, num_unknown)    ~ 1/num_pair of the above
-        residual / temp                       ~ num_pair * dtype_bytes
-    Empirically, ~3x of Gw covers the rest. Use VRAM_SAFETY as the
-    headroom factor.
-    """
-    if not (HAS_TORCH and torch.cuda.is_available()):
-        return DEFAULT_CHUNK_SIZE
-    free_b, _ = torch.cuda.mem_get_info()
-    per_pixel_bytes = 3 * num_pair * num_unknown * dtype_bytes
-    n = int(VRAM_SAFETY * free_b / max(per_pixel_bytes, 1))
-    return max(1, n)
 
 
 def _solve_cholesky(G_dev, w_dev, y_dev):
     """Per-pixel weighted least-squares via normal equations + batched Cholesky.
 
-    For each pixel k with weighted system Gw_k @ x_k = yw_k where
-    Gw_k = diag(w_k) @ G and yw_k = w_k * y_k, solve the normal equations
-        (Gw_k^T @ Gw_k) x_k = Gw_k^T @ yw_k
-    via cuSolver-batched Cholesky. Compared to a gels (QR) path this
-    collapses ~num_unknown Householder iterations per pixel into a single
-    batched factorization, reducing per-chunk kernel launches from
-    ~num_pixel * num_unknown to ~5.
-
-    Rank-deficient pixels are detected via cholesky_ex info codes; their
-    factor is replaced with identity and right-hand-side with zeros so the
-    downstream cholesky_solve produces an all-zero solution for those
-    pixels and never propagates NaN/Inf.
+    Assembles the weighted system
+        ``Gw_k = diag(w_k) @ G``,  ``yw_k = w_k * y_k``
+    in one broadcast and dispatches to the shared
+    ``solve_normal_equations_batched`` core.
 
     Args:
         G_dev: (num_pair, num_unknown) design matrix on GPU.
@@ -109,23 +56,8 @@ def _solve_cholesky(G_dev, w_dev, y_dev):
         (n, num_unknown) per-pixel solution.
     """
     Gw = w_dev.t().unsqueeze(-1) * G_dev.unsqueeze(0)
-    yw = (w_dev * y_dev).t().unsqueeze(-1)
-    Gw_T = Gw.transpose(-1, -2)
-    N = Gw_T @ Gw
-    r = Gw_T @ yw
-
-    L, info = torch.linalg.cholesky_ex(N)
-    fail_mask = info != 0
-    if fail_mask.any():
-        n_fail = int(fail_mask.sum().item())
-        print(f'WARNING: {n_fail} rank-deficient pixel(s) in chunk; '
-              'setting solution to zero')
-        eye = torch.eye(N.shape[-1], device=L.device, dtype=L.dtype)
-        L = torch.where(fail_mask.view(-1, 1, 1), eye, L)
-        r = torch.where(fail_mask.view(-1, 1, 1), torch.zeros_like(r), r)
-
-    X = torch.cholesky_solve(r, L)
-    return X.squeeze(-1)
+    yw = (w_dev * y_dev).t()
+    return solve_normal_equations_batched(Gw, yw)
 
 
 def estimate_timeseries_batch(
@@ -182,7 +114,7 @@ def estimate_timeseries_batch(
         raise ValueError(
             f"unsupported solver={solver!r}; choose from {SUPPORTED_SOLVERS}"
         )
-    device = _get_torch_device(solver)
+    device = get_torch_device(solver)
 
     G = B if min_norm_velocity else A
     num_pair, num_unknown = G.shape
@@ -201,7 +133,7 @@ def estimate_timeseries_batch(
 
     # decide chunk size
     if chunk_size is None or chunk_size <= 0:
-        chunk_size = _auto_chunk_size(num_pair, num_unknown)
+        chunk_size = auto_chunk_size(num_pair, num_unknown)
         if print_msg:
             free_gib = torch.cuda.mem_get_info()[0] / 2**30
             print(f'GPU auto chunk_size = {chunk_size} pixels '
